pyproject.toml
[project]
name = "executive-frequency"
version = "0.1.0"
description = "Scraper finding CEO/founder engineering problems on HN and Reddit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "praw>=7.7.1",
    "click>=8.1.0",
    "python-dateutil>=2.9.0",
]

[project.scripts]
execfreq = "main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "main.py",
    "hn_scraper.py",
    "reddit_scraper.py",
    "scorer.py",
]
